+++ src/com/hermant/program/instruction/LoadByteInstruction.java
@@ -12,7 +12,7 @@ public class LoadByteInstruction extends Instruction {
     public boolean execute(Machine m, boolean debug){
         super.execute(m, debug);
         int ramAddress = getMemoryAddress(m.getRegister());
-        m.getRegister().setInteger(reg1, (byte)m.getRam().getByte(ramAddress));
+        m.getRegister().setInteger(reg1, m.getRam().getByte(ramAddress));
         return true;
     }
